requirements.txt

chromadb==0.3.26
duckdb==0.8.1
langchain==0.0.198
matplotlib==3.7.1
numpy==1.24.3
openai==0.27.8
pandas==2.0.2
Pillow==9.5.0
PyMuPDF==1.22.5
pypdf==3.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
seaborn==0.12.2
streamlit==1.23.1
tiktoken==0.4.0
urllib3==1.26.16